Moving blog details logic into service layer
