+++ src/controllers/blog.controller.ts
@@ -1,17 +1,12 @@
 import { Response, Request } from 'express';
 import type { UserTokenPayload } from '../services/authentication';
-import { Blog } from '../models/blog';
-import { Comment } from '../models/comment';
-import { User } from '../models/user';
-
+import { getBlogDetails } from '../services/blog.service';
 import {
   AddBlogPostPayload,
   BlogWithCommentsResponse,
   BlogPostResponse,
   BlogListAPIResponse,
 } from '../types/blog.type';
-import { buildFileUrl } from '../utils/fileUrlGenerator';
-import { BlogLike } from '../models/blogLike';
 import { ApiError } from '../utils/ApiError';
 import { addBlogPost, getBlogList } from '../services/blog.service';
 
@@ -59,66 +54,14 @@ export const handleAddBlogPost = async (
 };
 
 export const handleGetBlogDetails = async (
-  req: Request<{ id: string }, {}, {}>,
+  req: Request<{ id: string }>,
   res: Response<BlogWithCommentsResponse>,
 ) => {
-  try {
-    const blog = await Blog.findById(req.params.id).populate('createdBy');
-    if (!blog) {
-      throw new ApiError(404, 'Blog not found');
-    }
+  const data = await getBlogDetails(req, req.params.id, req.user?.id);
 
-    const comments = await Comment.find({ blogId: req.params.id })
-      .populate('createdBy')
-      .sort({ createdAt: -1 });
-
-    let isLikedByUser = false;
-    if (req.user) {
-      const result = await BlogLike.findOne({ blogId: req.params.id, userId: req.user.id });
-      isLikedByUser = result?.isLiked ?? false;
-    }
-
-    const user = blog.createdBy instanceof User ? blog.createdBy : null;
-    const sanitizedBlog = {
-      id: blog._id.toString(),
-      title: blog.title,
-      body: blog.body,
-      coverImageUrl: buildFileUrl(req, blog.coverImageUrl),
-      isLikedByUser: isLikedByUser,
-      totalLikes: blog.likeCount,
-      createdBy: {
-        name: user ? user.name : 'Unknown',
-        imageUrl: user
-          ? buildFileUrl(req, user.profileImageUrl)
-          : buildFileUrl(req, '/images/default.png'),
-      },
-      createdAt: blog.createdAt,
-    };
-    const sanitizedComments = comments.map((comment) => {
-      const user = comment.createdBy instanceof User ? comment.createdBy : null;
-      return {
-        id: comment._id.toString(),
-        content: comment.content,
-        createdBy: {
-          name: user ? user.name : 'Unknown',
-          imageUrl: user
-            ? buildFileUrl(req, user.profileImageUrl)
-            : buildFileUrl(req, '/images/default.png'),
-        },
-        createdAt: comment.createdAt,
-      };
-    });
-
-    return res.json({
-      message: 'Blog details fetched successfully',
-      success: true,
-      data: {
-        blog: sanitizedBlog,
-        comments: sanitizedComments,
-      },
-    });
-  } catch (error) {
-    console.error('Error fetching blog details:', error);
-    return res.status(500).json({ success: false, message: 'Internal server error' });
-  }
+  return res.json({
+    message: 'Blog details fetched successfully',
+    success: true,
+    data,
+  });
 };
