+++ src/services/blog.service.ts
@@ -1,9 +1,12 @@
-// services/blog.service.ts
 import { Blog } from '../models/blog';
 import { Request } from 'express';
 import { buildFileUrl } from '../utils/fileUrlGenerator';
 import { AddBlogPostPayload, BlogCreationResponse, BlogItem } from '../types/blog.type';
 import { UserTokenPayload } from './authentication';
+import { Comment } from '../models/comment';
+import { BlogLike } from '../models/blogLike';
+import { User } from '../models/user';
+import { ApiError } from '../utils/ApiError';
 
 export const getBlogList = async (req: Request): Promise<BlogItem[]> => {
   const blogs = await Blog.find({}).populate('createdBy').sort({ createdAt: -1 });
@@ -40,5 +43,57 @@ export const addBlogPost = async (
       },
       createdAt: blog.createdAt,
     },
+  };
+};
+
+export const getBlogDetails = async (req: Request, blogId: string, userId?: string) => {
+  const blog = await Blog.findById(blogId).populate('createdBy');
+  if (!blog) {
+    throw new ApiError(404, 'Blog not found');
+  }
+
+  const comments = await Comment.find({ blogId }).populate('createdBy').sort({ createdAt: -1 });
+
+  let isLikedByUser = false;
+  if (userId) {
+    const result = await BlogLike.findOne({ blogId, userId });
+    isLikedByUser = result?.isLiked ?? false;
+  }
+
+  const user = blog.createdBy instanceof User ? blog.createdBy : null;
+  const sanitizedBlog = {
+    id: blog._id.toString(),
+    title: blog.title,
+    body: blog.body,
+    coverImageUrl: buildFileUrl(req, blog.coverImageUrl),
+    isLikedByUser,
+    totalLikes: blog.likeCount,
+    createdBy: {
+      name: user ? user.name : 'Unknown',
+      imageUrl: user
+        ? buildFileUrl(req, user.profileImageUrl)
+        : buildFileUrl(req, '/images/default.png'),
+    },
+    createdAt: blog.createdAt,
+  };
+
+  const sanitizedComments = comments.map((comment) => {
+    const user = comment.createdBy instanceof User ? comment.createdBy : null;
+    return {
+      id: comment._id.toString(),
+      content: comment.content,
+      createdBy: {
+        name: user ? user.name : 'Unknown',
+        imageUrl: user
+          ? buildFileUrl(req, user.profileImageUrl)
+          : buildFileUrl(req, '/images/default.png'),
+      },
+      createdAt: comment.createdAt,
+    };
+  });
+
+  return {
+    blog: sanitizedBlog,
+    comments: sanitizedComments,
   };
 };
